:sparkles: 연습문제 7-25, 7-26 내부 클래스의 멤버변수 출력하기

+++ src/Ex20.java
@@ -37,7 +37,7 @@ class Rectangle extends Shape {
     double width = 0;
     double height = 0;
 
-    public Rectangle(int w, int h) {
+    public Rectangle(double w, double h) {
         this.width = w;
         this.height = h;
     }
